added final edits

+++ controller/productController.js
@@ -2,7 +2,7 @@ const Product = require('../model/product')
 const Favorite = require('../model/favorite')
 const errorHandler = require('../utils/errorHandler')
 
-class ProductController{
+class ProductController {
 
     async getNewProducts(req, res) {
         try {
@@ -30,10 +30,11 @@ class ProductController{
         }
     }
 
+    // у скольких пользователей товар в избранном
     async countFavoriteByProduct(req, res) {
         try {
             const amount = await Favorite.count({
-                where: {productId: req.params.productId}
+                where: { productId: req.params.productId }
             })
             res.status(200).json(amount)
         } catch (error) {
@@ -42,31 +43,34 @@ class ProductController{
     }
 
     async addFavorite(req, res) {
-        try{
+        try {
             const candidate = await Favorite.findOne(
-                {where:
-                    { userId: Number(req.query.userId),
-                    productId: Number(req.query.productId)}})
-            if(candidate){
+                {
+                    where:
+                    {
+                        userId: Number(req.query.userId),
+                        productId: Number(req.query.productId)
+                    }
+                })
+            if (candidate) {
                 res.status(401).json({
                     message: 'This product already in favorites'
                 })
-            }else{
+            } else {
                 const favorite = await Favorite.create({
                     userId: Number(req.query.userId),
                     productId: Number(req.query.productId)
                 })
                 res.status(201).json(favorite)
             }
-            } catch(error){
-                errorHandler(res, error)
-            }
+        } catch (error) {
+            errorHandler(res, error)
+        }
     }
 
-    //не работает
     async deleteFavorite(req, res) {
         try {
-            await Favorite.destroy({where: {id: req.params.id}})
+            await Favorite.destroy({ where: { id: req.params.id } })
             res.status(200).json({
                 message: 'Fav deleted.'
             })
@@ -75,37 +79,48 @@ class ProductController{
         }
     }
 
+    //список избранного для конкретного пользователя
     async getFavProductByUser(req, res) {
         try {
             const favs = await Favorite.findAll({
                 where: {
-                            userId: req.params.userId,
-                        },
+                    userId: req.params.userId,
+                },
                 attributes: {
-                    exclude: ['id','userId']
+                    exclude: ['id', 'userId']
                 }
             })
-            const favProducts = await Promise.all( 
-                favs.map( 
-                    async ({ productId }) => 
-                    await Product.findOne({ 
-                        where: { 
-                            id : productId 
-                        }}))
+            const favProducts = await Promise.all(
+                // для каждого элемента массива
+                favs.map(
+                    async ({ productId }) =>
+                        await Product.findOne({
+                            where: {
+                                id: productId
+                            }
+                        }))
             )
             res.status(200).json(favProducts)
-            
-                } catch (error) {
-                    errorHandler(res, error)
-                }
-            }
+
+        } catch (error) {
+            errorHandler(res, error)
+        }
+    }
 
     async getOneProduct(req, res) {
         try {
-                const product = await Product.findOne({
-                where : { id: req.params.id }
+            const product = await Product.findOne({
+                where: {
+                    id: req.params.id
+                }
             })
-            res.status(200).json(product)
+            if (product) {
+                res.status(200).json(product)
+            } else {
+                res.status(404).json({
+                    message: 'Product not found'
+                })
+            }
         } catch (error) {
             errorHandler(res, error)
         }
@@ -129,12 +144,37 @@ class ProductController{
         }
     }
 
+    // async deleteProduct(req, res) {
+    //     try {
+    //         await Product.destroy({where: {id: req.params.id}})
+    //         res.status(200).json({
+    //             message: 'Product successfully uninstalled'
+    //         })
+    //     } catch (error) {
+    //         errorHandler(res, error)
+    //     }
+    // }
+
     async deleteProduct(req, res) {
         try {
-            await Product.destroy({where: {id: req.params.id}})
-            res.status(200).json({
-                message: 'Product deleted.'
+            const product = await Product.findOne({
+                where: {
+                    id: req.params.id
+                }
             })
+            if (product) {
+                await Favorite.destroy({
+                    where: { userId: req.params.id }
+                })
+                await Product.destroy({ where: { id: req.params.id } })
+                res.status(200).json({
+                    message: 'Product removed successfully'
+                })
+            } else {
+                res.status(404).json({
+                    message: 'Product not found'
+                })
+            }
         } catch (error) {
             errorHandler(res, error)
         }
@@ -143,12 +183,12 @@ class ProductController{
     async updateProducts(req, res) {
         try {
             await Product.update(req.body, {
-                where: {id: req.params.id}
+                where: { id: req.params.id }
             })
             res.status(200).json(
                 await Product.findOne({
-                where: {id: req.params.id}
-            }))
+                    where: { id: req.params.id }
+                }))
         } catch (error) {
             errorHandler(res, error)
         }
